Get all of the tests to build
