+++ src/test/datascience/testNativeEditorProvider.ts
@@ -4,14 +4,14 @@
 import { inject, injectable } from 'inversify';
 import { Event, Uri } from 'vscode';
 
-import { ICommandManager, IDocumentManager, IWorkspaceService } from '../../client/common/application/types';
+import { ICommandManager, ICustomEditorService, IWorkspaceService } from '../../client/common/application/types';
 import { IFileSystem } from '../../client/common/platform/types';
 import { IAsyncDisposableRegistry, IConfigurationService, IDisposableRegistry } from '../../client/common/types';
 import { InteractiveWindowMessageListener } from '../../client/datascience/interactive-common/interactiveWindowMessageListener';
 import { InteractiveWindowMessages } from '../../client/datascience/interactive-common/interactiveWindowTypes';
 import { NativeEditor } from '../../client/datascience/interactive-ipynb/nativeEditor';
 import { NativeEditorProvider } from '../../client/datascience/interactive-ipynb/nativeEditorProvider';
-import { IDataScienceErrorHandler, INotebookEditor, INotebookEditorProvider, INotebookServerOptions } from '../../client/datascience/types';
+import { INotebookEditor, INotebookEditorProvider, INotebookServerOptions } from '../../client/datascience/types';
 import { IServiceContainer } from '../../client/ioc/types';
 
 @injectable()
@@ -31,21 +31,10 @@ export class TestNativeEditorProvider implements INotebookEditorProvider {
         @inject(IWorkspaceService) workspace: IWorkspaceService,
         @inject(IConfigurationService) configuration: IConfigurationService,
         @inject(IFileSystem) fileSystem: IFileSystem,
-        @inject(IDocumentManager) documentManager: IDocumentManager,
         @inject(ICommandManager) cmdManager: ICommandManager,
-        @inject(IDataScienceErrorHandler) dataScienceErrorHandler: IDataScienceErrorHandler
+        @inject(ICustomEditorService) customEditorService: ICustomEditorService
     ) {
-        this.realProvider = new NativeEditorProvider(
-            serviceContainer,
-            asyncRegistry,
-            disposables,
-            workspace,
-            configuration,
-            fileSystem,
-            documentManager,
-            cmdManager,
-            dataScienceErrorHandler
-        );
+        this.realProvider = new NativeEditorProvider(serviceContainer, asyncRegistry, disposables, workspace, configuration, fileSystem, cmdManager, customEditorService);
     }
 
     public get activeEditor(): INotebookEditor | undefined {
@@ -56,8 +45,8 @@ export class TestNativeEditorProvider implements INotebookEditorProvider {
         return this.realProvider.editors;
     }
 
-    public async open(file: Uri, contents: string): Promise<INotebookEditor> {
-        const result = await this.realProvider.open(file, contents);
+    public async open(file: Uri): Promise<INotebookEditor> {
+        const result = await this.realProvider.open(file);
 
         // During testing the MainPanel sends the init message before our interactive window is created.
         // Pretend like it's happening now
